fix(ingredients): lift free-text resolver into shared module, adopt in UnifiedIngredientService
The conservative free-text → catalog resolver added in #555 lived inside
`ingredientNutritionAggregation.ts`, so only the recipe-nutrition aggregator
benefited. `UnifiedIngredientService.getIngredientByName` — which backs the
ESMS `matchRate` in `hierarchicalRecipeCalculations.ts` (called per recipe
ingredient) — stayed exact-match-only and resolved well under half of recipe
ingredient occurrences.

- Extract the resolver (strip qty/units, "X or Y"→first, drop prep adjectives,
  singularize incl -oes/-ves/irregulars, exact-normalized then whole-word
  token-subset) into `src/utils/ingredientResolution.ts`.
- `ingredientNutritionAggregation.ts` now imports + re-exports it (no behavior
  change; existing importers/tests untouched).
- `UnifiedIngredientService.getIngredientByName` falls back to the shared
  resolver only after its exact-match loop misses — exact still wins, so this
  only ever turns a previous miss into a hit (monotonic).
- `normName` now treats underscores as spaces, so the 17 catalog entries whose
  `name` carries the slug form ("rice_vinegar", "apple_cider_vinegar") finally
  resolve from their human phrase. Verified: "rice vinegar" was MISSING, now
  resolves; over-match guard holds ("rice" ≠ "ice").
- New `ingredientResolution.test.ts` (over-match guards + underscore case);
  all 17 resolver tests pass; typecheck + lint clean.
